--- frontend/src/app/groups/[id]/post/[postId]/edit/page.tsx
@@ -1,14 +1,0 @@
-import RequireAuthenticated from "@/lib/auth/components/RequireAuthenticated";
-import PostEditPage from "./EditPage";
-
-function Page() {
-  return (
-    <div className="container mx-auto p-4">
-      <RequireAuthenticated>
-        <PostEditPage />
-      </RequireAuthenticated>
-    </div>
-  );
-}
-
-export default Page;
