--- frontend/src/app/groups/[id]/post/page.tsx
@@ -1,14 +1,0 @@
-import RequireAuthenticated from "@/lib/auth/components/RequireAuthenticated";
-import PostListPage from "./PostList";
-
-function Page() {
-  return (
-    <div className="container mx-auto p-4">
-      <RequireAuthenticated>
-        <PostListPage />
-      </RequireAuthenticated>
-    </div>
-  );
-}
-
-export default Page;
